requirements.txt
flask
spacy
scikit-learn
PyPDF2
python-docx
reportlab
gunicorn
reportlab